+++ tests/Fublisher.Faker.AspNetCoreWindowService/Program.fs
@@ -1,0 +1,143 @@
+module CellScript.Server.Fcs.App
+
+open System
+open System.IO
+open Microsoft.AspNetCore.Builder
+open Microsoft.AspNetCore.Cors.Infrastructure
+open Microsoft.AspNetCore.Hosting
+open Microsoft.Extensions.Logging
+open Microsoft.Extensions.DependencyInjection
+open Giraffe
+open System.Diagnostics
+open System.Reflection
+open Microsoft.AspNetCore.Hosting.WindowsServices
+
+// ---------------------------------
+// Models
+// ---------------------------------
+
+type Message =
+    {
+        Text : string
+    }
+
+// ---------------------------------
+// Views
+// ---------------------------------
+
+module Views =
+    open GiraffeViewEngine
+
+    let layout (content: XmlNode list) =
+        html [] [
+            head [] [
+                title []  [ encodedText "CellScript.Server.Fcs" ]
+                link [ _rel  "stylesheet"
+                       _type "text/css"
+                       _href "/main.css" ]
+            ]
+            body [] content
+        ]
+
+    let partial () =
+        h1 [] [ encodedText "CellScript.Server.Fcs" ]
+
+    let index (model : Message) =
+        [
+            partial()
+            p [] [ encodedText model.Text ]
+        ] |> layout
+
+// ---------------------------------
+// Web app
+// ---------------------------------
+
+let indexHandler (name : string) =
+    let greetings = sprintf "Hello %s, from Giraffe!" name
+    let model     = { Text = greetings }
+    let view      = Views.index model
+    htmlView view
+
+let webApp =
+    choose [
+        GET >=>
+            choose [
+                route "/" >=> indexHandler "world"
+                routef "/hello/%s" indexHandler
+            ]
+        setStatusCode 404 >=> text "Not Found" ]
+
+// ---------------------------------
+// Error handler
+// ---------------------------------
+
+let errorHandler (ex : Exception) (logger : ILogger) =
+    logger.LogError(ex, "An unhandled exception has occurred while executing the request.")
+    clearResponse >=> setStatusCode 500 >=> text ex.Message
+
+// ---------------------------------
+// Config and Main
+// ---------------------------------
+
+let configureCors (builder : CorsPolicyBuilder) =
+    builder.WithOrigins("http://localhost:8080")
+           .AllowAnyMethod()
+           .AllowAnyHeader()
+           |> ignore
+
+let configureApp (app : IApplicationBuilder) =
+    let env = app.ApplicationServices.GetService<IHostingEnvironment>()
+    (match env.IsDevelopment() with
+    | true  -> app.UseDeveloperExceptionPage()
+    | false -> app.UseGiraffeErrorHandler errorHandler)
+        .UseHttpsRedirection()
+        .UseCors(configureCors)
+        .UseStaticFiles()
+        .UseGiraffe(webApp)
+
+let configureServices (services : IServiceCollection) =
+    services.AddCors()    |> ignore
+    services.AddGiraffe() |> ignore
+
+let configureLogging (builder : ILoggingBuilder) =
+    builder.AddFilter(fun l -> l.Equals LogLevel.Error)
+           .AddConsole()
+           .AddDebug() |> ignore
+
+[<EntryPoint>]
+let main args =
+
+    let isConsole =
+        (Debugger.IsAttached || Array.contains "--console" args)
+
+    let builder =
+        WebHostBuilder()
+            .UseKestrel()
+            .UseIISIntegration()
+            .Configure(Action<IApplicationBuilder> configureApp)
+            .ConfigureServices(configureServices)
+            .ConfigureLogging(configureLogging)
+
+    if isConsole
+    then 
+        let dir = 
+            let pathToExe = Process.GetCurrentProcess().MainModule.FileName
+            Path.GetDirectoryName pathToExe
+        Directory.SetCurrentDirectory dir
+
+
+    let contentRoot = Directory.GetCurrentDirectory()
+    let webRoot     = Path.Combine(contentRoot, "WebRoot")
+
+    let builder =
+        builder
+            .UseContentRoot(contentRoot)
+            .UseWebRoot(webRoot)
+            .Build()
+
+
+    if isConsole
+    then builder.Run()
+    else builder.RunAsService()
+
+    0