+++ RELEASE_NOTES.md
@@ -1,6 +1,9 @@
 ﻿# Release Notes
 
-## 0.1.27-alpha - tbd
+## 0.1.27 - tbd
+* Collaborators.NextRelease Target: check (github_token,github_releaseUser environments) first
+* Zip AspNetCore projects
+* Added FPublishe.Utils.Cli
 
 ## 0.1.26 - 2019-03-24
 * cancel dotnet pack --no-build